Más cambios en Parking

+++ src/trim02/parking/Coche.java
@@ -30,7 +30,7 @@ public class Coche extends Vehiculo {
 	}
 
 	@Override
-	public float calcularTarifa() {
+	public int calcularTarifa() {
 		int incremento = 0;
 		
 		if(this.longitud >= 2.5f && this.longitud < 3.5f) {
