+++ src/trim02/parking/Tarifable.java
@@ -1,5 +1,5 @@
 package trim02.parking;
 
 public interface Tarifable {
-	float calcularTarifa();
+	int calcularTarifa();
 }
